I figure out. I used something similar to factory method. I need to add the case of list page, and i will complete the factory.py
I decided to use the old model - models.py with minor changes and improvements, like decompositions and decorator to get a list instead of a dictionary to create something similar to data mapping.
Because of dynamic typing of python i don't need to write all that patterns (rewrite from the book), i can just take a main idea of these patterns and implement them.